Move JobFiltering component to queues folder
